Shutdown client when sbt exits
Presently the thin client stays open even when sbt has exited. This
commit changes the behavior so that before shutting down, sbt sends a
simple message to its connected clients instructing them to shutdown. In
order to handle graceful shutdown, I also moved the LineReader.readLine
off of the main thread to ensure that the main thread can be interrupted.