Use sbt source file stamps in zinc
This fixes the nio/external-hooks test and also restores the performance
of the benchmarks for the latest sbt version in
https://github.com/eatkins/scala-build-watch-performance which had
regressed when the custom ExternalHooks were disabled in
7c4b01d9f722252d3e7379f69591861bdd57152c.

The main change is that it changes the ReadStamps object that is passed
into the compiler options to one that uses the unmanagedFileStampCache
and managedFileStampCache for source files and falls back to the default
stamper otherwise. This improves the performance quite significantly
since we only hash the files once. It also makes it so that the analysis
file will contain the source file stamps of the files when compilation
began, rather than when compilation ended. That is what
nio/external-hooks was testing. In the real world what could happen was
that one modified a source file during compilation but then no
incremental re-compilation would occur because after the initial
compilation completed, zinc wrote the stamp of the modified source file
in the analysis file even though it may have actually compiled a
different version of the source file.

I also had to call toRealPath on the paths generated in rootPaths
because of some mac specific scripted issues. On mac, the location of
the user temp directory is something like /var/folders/7q/...
but is actually a symlink to /private/var/folders/7q/... The file stamps
generated with the nio apis resolve that symlink so there was a mismatch
between the paths generated by the inputFileStamps task and the result
of converting a virtual file to a regular path.

We probably could delete the `ExternalHooks` object entirely. The only
concern that I have is that it had some additional logic for forcing an
invalidation if