Clear appenders after task evaluation
sbt was leaking ConsoleAppenders with each task run. When I looked at a
heap dump after I ran compile 30 times in a project (see below) with
about 20 subprojects, I found the dump filled with leaked
ConsoleAppenders. This can be fixed by removing the loggers created
during task evaluation from log4j. I was concerned about removing
loggers that were still in use so I refactored LogManager so that it now
takes a LogExchange parameter. EvaluateTask can pass in a fresh
LogExchange and clear out all of the loggers when it's done.

Clearing the appenders is not only a memory leak but a performance
improvement as well. Adding a logger requires updating a concurrent hash
map which gets slower as the map size increases. I noticed in a
flamegraph that creating appenders was taking 50% of the total time
which was what motivated this change.

My project was a local clone of the repro project in
https://github.com/sbt/sbt/issues/5508. I was testing sbt overhead by
timing how long it took to run 30 iterations of no-op compile. There is
also a memory leak in TaskProgress that I will resolve separately that
causes degraded performance. With the fix for TaskProgress, the console
appender leaks cause the time for no-op compilation to increase per
iteration by about 100-200ms every 30 iterations. After applying this
change on top, there is no longer a noticeable increase in as the number
of iterations increase. This is very important since if we're planning
to push people towards the thin client, we really need to avoid resource
leaks in sbt.

This is not scientific, but for a sense of the potential impact of these
changes, with 1.3.13, the first 30 no-op compiles took about 390ms per
compile, the second 30 took 460 and the third took about 570. After all
of the fixes were applied in 1.4.0, the first 30 iteration took about
350ms per compile, the second took around 270. Then I ran 300
iterations which almost surely would have caused an OOM and taken
forever with 1.3.13 and the average was 250 (yay, JIT!).

Fixes #4779 https://github.com/sbt/sbt/pull/4779