Don't block sbt in `run` and `console` tasks
Currently when the user invokes run or console, these tasks tie up the
sbt task evaluation loop until they complete. This could be problematic
if the user is running, say, a server or an interactive console in their
sbt shell while also using metals with bsp. Metals would be unable to
compile any source files while these tasks were blocking. We can fix
this by running these tasks in the background in such a way that they
block the ui thread for the console or thin client session that
initiated them but not block the task evaluation loop. This is only safe
to do if we copy the classpath for these tasks so that it isn't blown
away by other tasks. In the case of run and runMain, this was already
done. The console tasks had to be reworked to use the same bgJobService
infrastructure as the run tasks.

Since there is the possibility that this causes unforeseen issues, there
is a new setting key, runInBackground, that if set to false will revert
to the old way of evaluating the run and console tasks. We also will
only run these tasks in the background if there is a server running.
Otherwise there is no point. This makes this change significantly less
risky than it might have otherwise been.

The trickiest part of the implementation to figure out was how to ensure
that System.out and System.in worked correctly for the background tasks.
Previously since only one instance of run or console could be run at a
time, we could just set a global variable for the current terminal and
redirecty System.out and System.in to the terminal's input and output
streams. Instead what we do here is we set an inheritable threadlocal
value for the task terminal. Because it is inheritable, if the run task
spawns additional threads, they should still get the same value. To set
the threadlocal terminal value correctly, it was necessary to propagate
more information about the terminal that initiated the currently running
command.

The final piece of the puzzle is that there needed to be some way of
finishing run and console without actually reporting their results. The
way this is implemented is by adding a new kind of Prompt called
Blocked. If run or console is running in the background, it will set its
prompt to Blocked before exiting. MainLoop will then strip all of the
remaining commands for that terminal and store them in the instance of
Blocked. It will also defer reporting the results. We also need to
modify Aggregation so that it does not print the timing results when run
and console exit since they haven't actually completed. A function that
prints the timing results is instead stored in the Blocked prompt and
printed when the task completes. When sbt returns to the main thread, it
reprompts the terminal that initiated the run or console task. Instead
of making an AskUserThread, we instead start a thread that blocks until
the run or console task completes. When it does, we print the timing
results and then re-enque whatever commands had been stripped by
MainLoop when it detected the blocked prompt. Reporting for the thin
client is handled by running a number of commands after the actual
command completes so this is a critical component to ensuring that the
thin client exits at all in batch mode and that it returns an error code
depending on the success or failure of the task.

Although the motivation of this change was to allow things like metals
compile in the background while a console is running, a side effect is
that multiple clients may run interactive console sessions
simultaneously on the same server.

This approach could be extended to test, but test is a less urgent case
because unlike run and console, it is extremely rare that test would
indefinitely block the task evaluation engine. Tests can take a while to
run, but at least they can be expected to exit in the vast majority of
use cases. At any rate, the same approach used here with the
bgJobService and BackgroundRunTask could be applied to the various test
tasks as well.